pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
dependency-injector = "^4.41.0"
python-ulid = "^2.7.0"
dynaconf = "^3.2.6"
psycopg2-binary = "^2.9.9"
passlib = "1.7.4"
pyjwt = "^2.9.0"
bcrypt = "3.2.0"
python-jose = "^3.3.0"

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["todo_app"]
preview = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "CPY001", "TID252"]

# [tool.ruff.lint.per-file-ignores]
# "__init__.py" = ["F401"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
